+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Upcoming
 ### Added
 - (Drupal) Concatenate arrays passed to render.
-- (Drupal) Add default variables as in template_preprocess().  
+- (Drupal) Add fallback Twig loader to emulate Drupal's handling of unscoped include/extend statements by looking for templates in a specified module/theme - [#88](https://github.com/LastCallMedia/Mannequin/issues/88)
+- (Drupal) Add default variables as in template_preprocess().
 
 ## 1.0.6
 ## 1.0.5